requirements.txt
pandas
requests
pyarrow
streamlit
plotly
python-dotenv